requirements.txt
tensorflow
numpy
Pillow
opencv-python